refactor: 엔티티 패키지 삭제 및 도메인 패키지로 통
