+++ backend/src/main/java/reviewme/highlight/domain/HighlightedLine.java
@@ -5,7 +5,6 @@ import java.util.Set;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import reviewme.highlight.domain.exception.HighlightIndexExceedLineLengthException;
-import reviewme.highlight.entity.HighlightRange;
 
 @Getter
 @EqualsAndHashCode
