+++ backend/src/main/java/reviewme/highlight/service/mapper/HighlightMapper.java
@@ -9,8 +9,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import reviewme.highlight.domain.HighlightedLines;
 import reviewme.highlight.domain.HighlightedLine;
-import reviewme.highlight.entity.Highlight;
-import reviewme.highlight.entity.HighlightRange;
+import reviewme.highlight.domain.Highlight;
+import reviewme.highlight.domain.HighlightRange;
 import reviewme.highlight.service.dto.HighlightIndexRangeRequest;
 import reviewme.highlight.service.dto.HighlightRequest;
 import reviewme.highlight.service.dto.HighlightedLineRequest;
