+++ backend/src/test/java/reviewme/highlight/entity/HighlightRangeTest.java
@@ -3,6 +3,7 @@ package reviewme.highlight.entity;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import org.junit.jupiter.api.Test;
+import reviewme.highlight.domain.HighlightRange;
 import reviewme.highlight.domain.exception.InvalidHighlightIndexRangeException;
 
 class HighlightRangeTest {
