+++ backend/src/test/java/reviewme/highlight/service/mapper/HighlightMapperTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import reviewme.fixture.QuestionFixture;
 import reviewme.fixture.ReviewGroupFixture;
-import reviewme.highlight.entity.Highlight;
-import reviewme.highlight.entity.HighlightRange;
+import reviewme.highlight.domain.Highlight;
+import reviewme.highlight.domain.HighlightRange;
 import reviewme.highlight.repository.HighlightRepository;
 import reviewme.highlight.service.dto.HighlightIndexRangeRequest;
 import reviewme.highlight.service.dto.HighlightRequest;
